add get review by id request

+++ internal/model/review/review.go
@@ -3,7 +3,7 @@ package review
 import "time"
 
 type Review struct {
-	ID        int64     `json:"id" validate:"required,min=1"`
+	ID        int       `json:"id" validate:"required,min=1"`
 	Author    string    `json:"author" validate:"required,min=1,max=255"`
 	Message   string    `json:"message" validate:"required"`
 	Rating    int       `json:"rating" validate:"required,min=0,max=5"`
